% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processing.R
\name{pre_process_bulk}
\alias{pre_process_bulk}
\title{Read and process featurecounts data in edgeR}
\usage{
pre_process_bulk(
  counts_filepath = NULL,
  sample_name_regex = "S[0-9]+",
  metadata = NULL,
  edger_min_count = 10,
  export_normalised_data = TRUE,
  export_pca = TRUE,
  pca_dims = c(5, 5),
  export_gene_boxplots = TRUE,
  boxplot_genes = NULL,
  boxplot_dims = c(4, 4),
  exclude_samples = NULL
)
}
\arguments{
\item{counts_filepath}{File path to the featurecounts output}

\item{sample_name_regex}{Regex for sample names to rename the columns}

\item{metadata}{Metadata file with sample information in tidy format. Columns should include
"sample" and "group" at minimum. "sample" column must correspond to "sample_name_regex",
and "group" column denotes the comparison groups i.e. stim or unstim.}

\item{edger_min_count}{Corresponds to the edgeR `min.count` function}

\item{export_normalised_data}{Should the normalised data be exported?}

\item{exclude_samples}{Samples to be excluded. Based on metadata$sample column}
}
\value{
edgeR object with norm factors calculated
}
\description{
This function takes an input from the featurecounts output,
cleans the expression file with better names, creates an edgeR object,
filters the data based on min.count, performs TMM-CPM normalisation,
and exports boxploits of the data distributions.
}
\examples{
\dontrun{
  data <- pre_process_bulk(counts_filepath = "counts.txt",
                           sample_name_regex = "S[0-9]+",
                           metadata = metadata,
                           edger_min_count = 10,
                           export_normalised_data = TRUE)
}

}
