% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre-processing.R
\name{process_bulk}
\alias{process_bulk}
\title{Process and perform stats testing with edgeR}
\usage{
process_bulk(
  edger_object = NULL,
  metadata = NULL,
  stats_test = "glmqlf",
  contrast_matrix = NULL,
  export_degs = TRUE
)
}
\arguments{
\item{edger_object}{edgeR object with norm factors calculated}

\item{metadata}{Metadata for comparisons. metadata$group is used for the design matrix}

\item{stats_test}{What stats test to use. Currently only glmqlf}

\item{contrast_matrix}{Contrast matrix for comparisons. Created with the `makeContrasts()`
function. See https://github.com/jackbibby1/bulkPipeline for example}

\item{export_degs}{Should differential expression testing results be exported?}
}
\value{
edgeR object with norm factors calculated
}
\description{
This function takes an edgeR input (typically from `pre_process_bulk()`)
creates the design matrix, estimates dispersion, and performs DE testing
}
\examples{
\dontrun{
  data <- process_bulk(edger_object = y,
                       metadata = metadata,
                       contrast_matrix = contrast_matrix)
}

}
